fix(ci,stella-tui): close three gate holes — a ratchet that grandfathered, a step that could not run, and a file with no headroom
Three independent defects in the machinery that is supposed to stop bad
changes landing. Each had the same shape: the guard reported success while
establishing nothing.

check-typed-errors.py --update (#3750)
--------------------------------------
The writer spelled an absent crate's allowance `baseline.get(c, n)` — the
crate's own new count — so the comparison was `n > n`, permanently False.
The one case the ratchet exists to block, a crate introducing its FIRST
`Result<_, String>`, was therefore the one case `--update` grandfathered:
it wrote the entry, printed a success line, and exited 0. The check path
never had the bug; it has always spelled the default `0`.

Both comprehensions now default to 0, and `raised` stops indexing
`baseline[c]` directly — that would have raised KeyError the moment an
absent crate was correctly flagged. A crate with no entry is reported as
being *added*, not "raised from 0", because that names the remedy.
AGENTS.md:302 already claimed "make typed-errors-update errors rather
than write a bigger number"; this makes the sentence true.

scripts/test-typed-errors.sh is new: 19 hermetic cases over throwaway
workspace trees, in the shape of test-file-size.sh.

make shellcheck (#3615)
-----------------------
On a machine without the binary the recipe died with `shellcheck: not
found` and `make: *** Error 127`, which reads exactly like a lint finding
in a script somebody just edited. Every rung of the gate ladder includes
this step, so no agent or human in that image could run the gate at all.

It now fails with an explicit "UNAVAILABLE — THIS STEP DID NOT RUN"
notice naming the install command. It fails rather than skipping,
because a gate step that can no-op is a gate step that will. The
distinction is carried by the MESSAGE, not the exit code: GNU make
normalises any recipe failure to its own status 2, so no caller can read
"could not run" out of the number. The lint invocation is byte-identical
when shellcheck is present.

scripts/test-shellcheck-guard.sh is new: 7 hermetic cases over mirrored
PATHs, pinning both halves — the notice on an absent binary, and that a
present binary is still invoked with argv intact and its findings still
fatal.

This is #3615's option 2. Its preferred option 1 — install shellcheck in
the dev container image — is not reachable from this repository: there is
no .devcontainer/, and the four Dockerfiles present build stella-serve,
the arenabench recorder, the arenabench runner and a bench sentinel, none
of which is the image agents run in. Filed as #3830 so the preference
does not evaporate when #3615 closes.

setup-dev-env.sh, three defects found while adding its shellcheck row:
  - the `ci` tier printed a red `miss` line and did not touch the exit
    code, so `--check` reported "all required tools present" and exited 0
    with shellcheck missing — the same red-line-green-verdict shape this
    commit is otherwise about. It is now fatal, which also makes the
    documented `--check` contract true for cargo-deny, silently
    non-fatal until now.
  - the workspace sniff tested `$repo_root/stella-core`, a path that
    stopped existing when the crates moved under `crates/`. The whole
    script had been exiting 1 on every tree.
  - the "make gate vs CI" heredoc carried a hand-written nine-step list
    against the real twenty-nine, omitting `shellcheck` itself. It was a
    third copy of GATE_STEPS that check-gate-parity.sh never saw, rotted
    in the under-reporting direction #1437 warns about. It is now derived
    from `make print-gate-steps`.

crates/stella-tui/src/envelope.rs (#3493)
-----------------------------------------
1371 lines against a flat 1500-line limit, not grandfathered, and the
file every envelope-contract change has to touch. (#3493 says 1496; that
figure does not reproduce here and `df89681` is unreachable from this
shallow clone, so it is recorded as unverified in a correction on the
issue rather than quietly restated.)

Three cohesive clusters move out to submodules beside the existing
mcp.rs / roles.rs / steering.rs, re-exported so every
`stella_tui::envelope::…` path and the crate-root `pub use` list are
unchanged: engine_config.rs, skills.rs, tool_policy.rs. 1371 -> 1002
lines. Nothing outside envelope.rs and envelope/ is touched, lib.rs is
byte-identical, and no baseline entry was added.

Verification
------------
test-typed-errors.sh: 19 passed / 0 failed; 16 passed / 3 failed against
`git show dfa35fd:scripts/check-typed-errors.py` — the fail->pass flip.
test-shellcheck-guard.sh: 7 passed / 0 failed; 3 passed / 4 failed
against `git show dfa35fd:Makefile`, reproducing the literal `Error 127`.
setup-dev-env.sh --check with only shellcheck hidden from PATH: exit 1
(was exit 0).
make guards-fast: exit 0, all 29 steps.
cargo fmt --check, cargo clippy -p stella-tui --all-targets -D warnings,
cargo test -p stella-tui: green.
Both ratchet baselines are untouched (`git diff dfa35fd --` on each is
empty).

The envelope split is a pure move and ships no witness, as #3493 states
it should not.

Closes #3750
Closes #3615
Closes #3493
Refs #2949, #3820, #3830
