+++ crates/stella-tui/src/envelope.rs
@@ -890,393 +890,24 @@ pub enum WorkspaceInput {
     Quit,
 }
 
-/// Which pipeline agent a per-agent engine override applies to — the
-/// configurable "agents" of `agent_engine_config`. `Default` is the
-/// interactive/step-loop agent; the rest are the staged pipeline's roles.
-///
-/// A mirror of `stella-cli`'s `EngineAgentKind`, in the same order, so
-/// `EngineConfigState::agents[i]` is that enum's `ALL[i]` — the driver maps
-/// one onto the other rather than this crate depending on the settings engine.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum EngineRole {
-    Default,
-    Worker,
-    Verifier,
-    Triage,
-    Research,
-    Plan,
-}
-
-impl EngineRole {
-    /// Stable settings-key / display order.
-    pub const ALL: [EngineRole; 6] = [
-        EngineRole::Default,
-        EngineRole::Worker,
-        EngineRole::Verifier,
-        EngineRole::Triage,
-        EngineRole::Research,
-        EngineRole::Plan,
-    ];
-
-    /// The `agent_engine_config.agents.<key>` settings key (and display
-    /// label) for this agent.
-    pub fn key(self) -> &'static str {
-        match self {
-            EngineRole::Default => "default",
-            EngineRole::Worker => "worker",
-            EngineRole::Verifier => "verifier",
-            EngineRole::Triage => "triage",
-            EngineRole::Research => "research",
-            EngineRole::Plan => "plan",
-        }
-    }
-}
-
-/// One agent's engine overrides as the ENGINE overlay edits them — a
-/// TUI-local mirror of `stella-cli`'s `agent_engine_config` per-agent
-/// settings, decoupled (like [`InstalledAgentEntry`]) so the TUI crate
-/// stays independent of the settings engine; the driver maps one to the
-/// other. Every field is optional — `None` renders as "provider default"
-/// and is omitted from the saved JSON (the screenshot's unchecked
-/// "Include" box).
-#[derive(Clone, Debug, Default, PartialEq)]
-pub struct EngineAgentState {
-    /// Model slug — `"provider/slug"` or a bare slug.
-    pub model: Option<String>,
-    /// Gateway/provider id override (`"anthropic"`, `"openrouter"`,
-    /// `"zai"`, or a settings-defined provider). When set, `model` is sent
-    /// verbatim as the slug to THIS provider — how an OpenRouter key routes
-    /// an `openai/...` slug.
-    pub provider: Option<String>,
-    /// Custom system prompt replacing the built-in one for this agent.
-    pub prompt: Option<String>,
-    /// Reasoning effort: `low|medium|high|xhigh|max`.
-    pub effort: Option<String>,
-    /// Thinking mode on/off (`None` = provider default).
-    pub reasoning: Option<bool>,
-    pub temperature: Option<f32>,
-    pub top_p: Option<f32>,
-    pub top_k: Option<u32>,
-    pub frequency_penalty: Option<f32>,
-    pub presence_penalty: Option<f32>,
-    pub repetition_penalty: Option<f32>,
-    pub max_tokens: Option<u32>,
-    pub seed: Option<u64>,
-    /// `low|medium|high`.
-    pub verbosity: Option<String>,
-    /// `auto|default|flex|priority`.
-    pub service_tier: Option<String>,
-}
-
-/// One engine role's **resolved** wiring — what that role will actually send,
-/// and the setting that decided it. The `/models` dialog's row type.
-///
-/// Every cell arrives pre-rendered from the driver, and deliberately so. The
-/// resolution is a precedence chain (`agents.<role>.model` over the flat
-/// `pipeline_<role>_model` over `default_model` over the session pin, with
-/// `--model` owning the worker and `effort_auto`/`reasoning_auto` able to
-/// replace a pin outright), it lives beside the request path in
-/// `stella-cli`'s `config_wiring`, and `stella config` already prints these
-/// exact cells. Re-deriving any of it here would give the deck a second
-/// answer that can drift from the engine's — and a routing report that can
-/// disagree with what runs is worse than no report at all.
-///
-/// This is why the fields are strings and not a typed mirror of the settings:
-/// the type carries a *rendering*, not a model to reason over.
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct RoleWiringRow {
-    /// The settings key this role is spelled as (`default` / `worker` /
-    /// `verifier` / `triage`) — the word a user would type to change it.
-    pub role: String,
-    /// `provider/slug` exactly as it goes on the wire.
-    pub model: String,
-    /// The resolved reasoning effort, or `provider default`. Names the pinned
-    /// value `effort_auto` discarded when it discarded one.
-    pub effort: String,
-    /// `thinking on` / `thinking off` / `thinking default`, with the same
-    /// `reasoning_auto` disclosure.
-    pub thinking: String,
-    /// The settings key that decided `model`, as a path a user can edit:
-    /// `agents.verifier.model`, `pipeline_triage_model`, `default_model`,
-    /// `--model (this invocation)`, or `session default`.
-    pub source: String,
-    /// What a session started **now** would resolve for this role, when a
-    /// saved settings edit makes that differ from the four cells above —
-    /// `None` when the saved settings agree with what is running, which is the
-    /// ordinary case.
-    ///
-    /// The cells above stay the session-start resolution, because "what is
-    /// running" is the question this dialog exists to answer and showing a
-    /// mid-session edit as though it were in force would misreport exactly
-    /// that. But saying nothing about the edit was its own lie: a user who
-    /// changed `pipeline_verifier_model` and saved saw their old pin with no
-    /// explanation, and read the dialog as having ignored the save (#1521).
-    /// So this is strictly *additional* information, pre-rendered driver-side
-    /// like every other cell, holding only the parts that differ.
-    pub next_session: Option<String>,
-}
-
-/// The full agent-engine configuration snapshot the ENGINE overlay renders
-/// and edits ([`Inbound::EngineConfig`] / [`WorkspaceInput::EngineConfigSave`]).
-#[derive(Clone, Debug, Default, PartialEq)]
-pub struct EngineConfigState {
-    /// Auto verifier-model selection: pick the best allowed model for the
-    /// verifier (preferring a different family than the worker's).
-    pub auto_mode: bool,
-    /// Auto per-agent effort (verifier high, worker medium, triage low).
-    pub effort_auto: bool,
-    /// Auto per-agent reasoning (on for verifier/worker, off for triage).
-    pub reasoning_auto: bool,
-    /// The model slugs the model pickers offer (`allowed_models`). Empty
-    /// means "no restriction" — pickers fall back to `catalog_models`.
-    pub allowed_models: Vec<String>,
-    /// Every configured provider id, for the provider picker.
-    pub providers: Vec<String>,
-    /// `provider/slug` strings from the catalog, scoped by the driver to
-    /// providers with a usable credential — the picker's fallback
-    /// vocabulary when `allowed_models` is empty.
-    pub catalog_models: Vec<String>,
-    /// Per-model effort vocabularies, keyed by the same `provider/slug`
-    /// strings (plus any `allowed_models` spec): the effort levels this
-    /// model, as served by this provider, can actually act on. An empty
-    /// list means effort is not a knob for that model (no reasoning
-    /// support, or an on/off-only thinking switch); a model absent from
-    /// the map is unknown and keeps the full vocabulary.
-    pub model_efforts: std::collections::HashMap<String, Vec<String>>,
-    /// Exactly one entry per [`EngineRole::ALL`] slot, in that order.
-    pub agents: Vec<EngineAgentState>,
-    /// What each role **resolves to**, which is a different question from
-    /// `agents` above: those are the raw overrides the ENGINE overlay edits,
-    /// where `None` means "inherit", and these are the answers after the
-    /// precedence chain and the auto-modes have run. The `/models` dialog
-    /// renders these; the overlay's editors never touch them.
-    ///
-    /// **An open vocabulary** (#3472), unlike `agents` above: the roles the
-    /// deck knows, then any role the host contributed, in the order it sent
-    /// them — see [`roles`] for the fold and what stays closed. Empty on a
-    /// driver that sent no wiring; the dialog says so rather than inventing
-    /// rows.
-    pub roles: Vec<RoleWiringRow>,
-}
-
-impl EngineConfigState {
-    /// The state for `role`, if present (the driver always sends all four).
-    pub fn agent(&self, role: EngineRole) -> Option<&EngineAgentState> {
-        EngineRole::ALL
-            .iter()
-            .position(|r| *r == role)
-            .and_then(|i| self.agents.get(i))
-    }
-}
-
-/// Which settings file carries the `"tools"` key that switched a tool off.
-///
-/// Not the same list as [`AgentScope`], and deliberately so: the editor writes
-/// to two scopes but has to *report* three, because the third one is the one
-/// it cannot write.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ToolScope {
-    /// The org-managed settings file. A denial here is a ceiling — no user or
-    /// project switch can lift it, which is why such a row renders locked.
-    Managed,
-    /// `~/.stella/settings.json`.
-    User,
-    /// `<workspace>/.stella/settings.json`.
-    Project,
-}
-
-impl ToolScope {
-    pub fn label(self) -> &'static str {
-        match self {
-            ToolScope::Managed => "org-managed",
-            ToolScope::User => "user",
-            ToolScope::Project => "project",
-        }
-    }
-}
-
-/// Why one tool is off: the `"tools"` key that did it, and where that key
-/// lives. Both halves matter — `"off"` alone leaves an operator hunting three
-/// settings files for a switch they may not have written themselves, and the
-/// key alone does not say which file to edit.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ToolDenial {
-    /// The exact tool name, its group (`"process"`), or `"*"` — resolved
-    /// most-specific-first, the same order the policy itself resolves.
-    pub key: String,
-    /// The scope whose file carries that key. `None` when the merged posture
-    /// and the individual scopes disagree (files changed under the snapshot) —
-    /// the row still reports the key rather than inventing a scope.
-    pub scope: Option<ToolScope>,
-}
-
-/// One tool the session actually has, as the SETTINGS tab's tool editor lists
-/// it. Driver-built (only the driver can see the assembled tool stack — MCP
-/// servers and a customer's own registered tools exist nowhere else), and
-/// deliberately NOT carrying an `enabled` flag: on/off is derived from
-/// [`ToolPolicyState::switches`] plus the panel's unsaved edits, so there is
-/// exactly one answer to "is this on?" rather than two that can drift.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ToolRow {
-    /// The exact dispatch name — and the exact settings key toggling this row
-    /// writes.
-    pub name: String,
-    /// The family this tool belongs to (`"file"`, `"process"`, …, or `"mcp"` /
-    /// `"custom"` for tools the catalog has never heard of). The editor's
-    /// section headers, and the key a header toggle writes.
-    pub group: String,
-    /// The org-managed scope denies this tool. The row is read-only: it cannot
-    /// be toggled on, and it must not render as though it could.
-    pub locked: bool,
-    /// Why it is off under the SAVED settings, or `None` when it is on.
-    /// Unsaved edits in the panel do not change this — it describes disk.
-    pub off: Option<ToolDenial>,
-}
-
-/// The tool-switch snapshot the SETTINGS tab's tool editor renders
-/// ([`Inbound::ToolPolicy`]).
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct ToolPolicyState {
-    /// Every tool this session can dispatch, before the policy filters it:
-    /// built-ins the environment satisfied, each connected MCP server's, and
-    /// each custom tool the workspace registered. Sorted by group then name.
-    pub tools: Vec<ToolRow>,
-    /// The merged `"tools"` section in force — tool names, group names, and
-    /// `"*"` mapped to on/off, with the org-managed ceiling already folded in.
-    /// The panel resolves a row's live value against this map (plus its own
-    /// unsaved edits) using the policy's own most-specific-first precedence.
-    pub switches: BTreeMap<String, bool>,
-}
-
-/// Which scope a skill lives in / is installed to. The loader reads both
-/// (`<workspace>/.stella/skills` and `~/.stella/skills`); the SKILLS
-/// tab asks this on install/create.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum SkillScope {
-    /// `<workspace>/.stella/skills` — travels with the repo.
-    Project,
-    /// `~/.stella/skills` — the user-global directory.
-    User,
-}
-
-impl SkillScope {
-    pub fn label(self) -> &'static str {
-        match self {
-            SkillScope::Project => "project",
-            SkillScope::User => "user",
-        }
-    }
-}
-
-/// One installed-skill row in the SKILLS tab — a driver snapshot, deliberately
-/// decoupled from `stella_core::skills::Skill` so the TUI crate stays
-/// independent of the skills engine (the driver maps one to the other).
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct SkillRow {
-    pub scope: SkillScope,
-    pub name: String,
-    pub description: String,
-    /// The pinned version's markdown body — carried so the tab can open the
-    /// edit overlay with no extra round-trip.
-    pub body: String,
-    /// Provenance label — `"workspace"`, `"user"`, `"installed"`, `"auto"`.
-    pub origin: String,
-    /// Session/persistent enabled state (a disabled skill is excluded from
-    /// recall injection; the file stays on disk).
-    pub enabled: bool,
-    /// The pinned/current version the recall path uses.
-    pub version: u32,
-    /// The highest version on disk (`version < latest` ⇒ pinned to an older one).
-    pub latest: u32,
-    /// Whether the deck may uninstall (delete) it.
-    pub removable: bool,
-}
-
-/// One registry search hit, parsed from `npx skills find` into structured
-/// fields — never the raw ANSI-laden line. `id` (`owner/repo@skill`) is the
-/// display name AND the token passed verbatim to `npx skills add` / `use`;
-/// `installs` is the human popularity string (`"15.8K installs"`, empty if the
-/// registry printed none) with `installs_rank` its numeric form for ranking/
-/// the popularity bar; `url` is the `skills.sh` page (shown in the preview).
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct SkillSearchHit {
-    pub id: String,
-    pub installs: String,
-    pub installs_rank: u64,
-    pub url: String,
-}
-
-/// The full installed-skills read-model rendered by the SKILLS tab.
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct SkillsView {
-    pub rows: Vec<SkillRow>,
-    /// A one-line status/outcome hint (last op result), or `None`.
-    pub status: Option<String>,
-    /// True while a driver op (npx search/install, LLM create) is in flight,
-    /// so the tab can show a working state.
-    pub busy: bool,
-    /// The name of the skill a just-completed [`SkillOp::Create`] wrote,
-    /// when this snapshot is that op's completion. The create dialog
-    /// transitions into the ctrl+o preview of exactly this row; `None` on a
-    /// failed create (the dialog shows `status` as the error) and on every
-    /// other snapshot.
-    pub created: Option<String>,
-}
-
-/// A SKILLS-tab operation routed to the driver, which owns the disk + npx +
-/// model. The deck emits one and folds back the refreshed [`Inbound::Skills`]
-/// / [`Inbound::SkillSearch`] snapshot.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum SkillOp {
-    /// Re-enumerate installed skills → [`Inbound::Skills`].
-    List,
-    /// Persistently enable/disable a skill → re-send the list.
-    SetEnabled {
-        scope: SkillScope,
-        name: String,
-        enabled: bool,
-    },
-    /// Delete a skill entirely from disk (deck double-confirms first).
-    Uninstall { scope: SkillScope, name: String },
-    /// `npx skills find <query>` → [`Inbound::SkillSearch`].
-    Search { query: String },
-    /// Fetch a search hit's `SKILL.md` for the ctrl+o preview overlay
-    /// (`npx skills use <id>`, extracting the wrapped body) →
-    /// [`Inbound::SkillPreview`]. No disk write — preview only.
-    Preview { id: String },
-    /// Install a registry skill into `scope` → refresh the list.
-    Install { scope: SkillScope, id: String },
-    /// LLM-assisted creation from a short description → refresh the list.
-    Create {
-        scope: SkillScope,
-        description: String,
-    },
-    /// Save an edited body as a NEW version (increments + pins it).
-    Edit {
-        scope: SkillScope,
-        name: String,
-        body: String,
-    },
-    /// Pin `version` as the active one WITHOUT editing (no version bump).
-    Pin {
-        scope: SkillScope,
-        name: String,
-        version: u32,
-    },
-}
-
 /// The `/models` role table: the open role vocabulary and its render order.
 pub mod roles;
 
-mod mcp;
+mod engine_config;
 /// The MCP tab's read models (rows, search results, inspector detail).
+mod mcp;
+mod skills;
 /// Why Esc steers rather than cancels — doc-only.
 pub mod steering;
+mod tool_policy;
+pub use engine_config::{EngineAgentState, EngineConfigState, EngineRole, RoleWiringRow};
 pub use mcp::{
     McpLiveIdentity, McpLookupState, McpSearchItem, McpSearchOutcome, McpServerDetail,
     McpServerInfo, McpToolRow,
 };
 pub use roles::{RoleTableEntry, role_table};
+pub use skills::{SkillOp, SkillRow, SkillScope, SkillSearchHit, SkillsView};
+pub use tool_policy::{ToolDenial, ToolPolicyState, ToolRow, ToolScope};
 
 /// A secret string whose `Debug` is redacted, so it can ride the deck's input
 /// channel (and any debug log of it) without leaking. The value is readable
